+++ BackupEngine/Backup/FullSaveStrategy.cs
@@ -1,14 +1,7 @@
-﻿using System;
-using System.IO;
-using System.Linq;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
+﻿using BackupEngine.Progress;
 using BackupEngine.State;
 using BackupEngine.Log;
-using BackupEngine.Settings;
 using System.Diagnostics;
-using System.Collections.Concurrent;
 
 namespace BackupEngine.Backup
 {
@@ -17,21 +10,6 @@ namespace BackupEngine.Backup
     /// </summary>
     public class FullSaveStrategy : SaveStrategy
     {
-        private SettingsRepository _settingsRepository = new SettingsRepository();
-        private readonly int _koLimit = 1024;
-        private readonly SemaphoreSlim _largeFileSemaphore = new SemaphoreSlim(1, 1);
-        /// <summary>
-        /// Mutex name to ensure only one CryptoSoft instance runs at a time.
-        /// </summary>
-        private static readonly string _mutexName = "Global\\CryptoSoft_Mutex";
-        /// <summary>
-        /// Queue to store files that require encryption.
-        /// </summary>
-        private readonly ConcurrentQueue<(string, string)> _cryptoQueue = new ConcurrentQueue<(string, string)>();
-        /// <summary>
-        /// Task responsible for processing the encryption queue.
-        /// </summary>
-        private Task _cryptoTask;
         /// <summary>
         /// Initializes a new instance of the FullSaveStrategy class.
         /// </summary>
@@ -39,7 +17,7 @@ namespace BackupEngine.Backup
         /// <summary>
         /// Executes the full backup process.
         /// </summary>
-        public override void Save(string uniqueDestinationPath)
+        public override void Save(string uniqueDestinationPath, EventWaitHandle waitHandle)
         {
             if (_configuration.ExtensionsToSave != null)
             {
@@ -63,10 +41,17 @@ namespace BackupEngine.Backup
             long remainingSize = totalSize;
 
             OnStateUpdated(new StateEvent("Full Backup", "Active", totalFiles, totalSize, remainingFiles, remainingSize, "", ""));
+            OnProgress(new ProgressEvent(
+                totalSize,
+                remainingSize
+            ));
+            
             List<Task> tasks = new List<Task>();
             WaitForBusinessSoftwareToClose();
+            
             foreach (string file in files)
             {
+                waitHandle.WaitOne();
                 string relativePath = file.Substring(sourcePath.Length + 1);
                 string destFile = Path.Combine(uniqueDestinationPath, relativePath);
                 Directory.CreateDirectory(Path.GetDirectoryName(destFile));
@@ -80,11 +65,11 @@ namespace BackupEngine.Backup
                     tasks.Add(Task.Run(() =>
                     {
                         WaitForBusinessSoftwareToClose(); 
-                        TransferFile(file, destFile, ref remainingFiles, ref remainingSize);
+                        TransferFile(file, destFile, ref totalSize, ref remainingFiles, ref remainingSize, ref waitHandle);
                     }));
                 }
             }
-
+            
             _cryptoTask = Task.Run(() =>
             {
                 WaitForBusinessSoftwareToClose(); 
@@ -95,13 +80,14 @@ namespace BackupEngine.Backup
             _cryptoTask.Wait();
 
             OnStateUpdated(new StateEvent("Full Backup", "Completed", totalFiles, totalSize, 0, 0, "", ""));
+            OnProgress(new ProgressEvent(totalSize,0));
             Console.WriteLine($"Full backup completed in: {uniqueDestinationPath}");
         }
 
         /// <summary>
         /// Transfers a file and updates the backup state.
         /// </summary>
-        private void TransferFile(string file, string destFile, ref int remainingFiles, ref long remainingSize)
+        private void TransferFile(string file, string destFile, ref long totalSize, ref int remainingFiles, ref long remainingSize, ref EventWaitHandle waitHandle)
         {
             FileInfo fileInfo = new FileInfo(file);
             bool isLargeFile = fileInfo.Length > _koLimit * 1024;
@@ -117,6 +103,9 @@ namespace BackupEngine.Backup
                     _largeFileSemaphore.Wait(); // Assure un seul fichier volumineux à la fois
                 }
 
+                // Check for pausing
+                waitHandle.WaitOne();
+
                 DateTime start = DateTime.Now;
                 TransferStrategy.TransferFile(file, destFile);
                 DateTime end = DateTime.Now;
@@ -126,7 +115,7 @@ namespace BackupEngine.Backup
                 remainingSize -= fileInfo.Length;
 
                 OnStateUpdated(new StateEvent("Full Backup", "Active", remainingFiles, remainingSize, remainingFiles, remainingSize, file, destFile));
-
+                OnProgress(new ProgressEvent(totalSize,remainingSize));
                 TransferEvent transferEvent = new TransferEvent(_configuration, duration, fileInfo, new FileInfo(destFile));
                 OnTransfer(transferEvent);
             }
@@ -140,77 +129,6 @@ namespace BackupEngine.Backup
                 if (isLargeFile)
                 {
                     _largeFileSemaphore.Release(); // Libère le sémaphore après transfert
-                }
-            }
-        }
-
-        /// <summary>
-        /// Waits for business software to close before proceeding with the backup.
-        /// </summary>
-        private void WaitForBusinessSoftwareToClose()
-        {
-            List<string> businessApps = _settingsRepository.GetBusinessSoftwareList();
-            while (IsBusinessSoftwareRunning(businessApps))
-            {
-                Console.WriteLine("Un logiciel métier est en cours d'exécution");
-                Thread.Sleep(3000);
-            }
-        }
-        /// <summary>
-        /// Checks if any business software is currently running.
-        /// </summary>  
-        private bool IsBusinessSoftwareRunning(List<string> businessApps)
-        {
-            foreach (var process in Process.GetProcesses())
-            {
-                if (businessApps.Contains(process.ProcessName, StringComparer.OrdinalIgnoreCase))
-                {
-                    return true;
-                }
-            }
-            return false;
-        }
-        /// <summary>
-        /// Determines if a file requires encryption based on its extension.
-        /// </summary>
-        private bool RequiresEncryption(string file)
-        {
-            string extension = Path.GetExtension(file);
-            if (_configuration.ExtensionsToSave == null)
-            {
-                return false;
-            }
-            return _configuration.ExtensionsToSave.Contains(extension);
-        }
-        /// <summary>
-        /// Processes the queue of files that need encryption.
-        /// </summary>
-        private void ProcessCryptoQueue()
-        {
-            using (Mutex mutex = new Mutex(false, _mutexName))
-            {
-                while (!_cryptoQueue.IsEmpty)
-                {
-                    if (_cryptoQueue.TryDequeue(out var filePair))
-                    {
-                        string source = filePair.Item1;
-                        string destination = filePair.Item2;
-
-                        if (!mutex.WaitOne(0, false))
-                        {
-                            Console.WriteLine("CryptoSoft est déjà en cours d'exécution");
-                            mutex.WaitOne();
-                        }
-
-                        try
-                        {
-                            TransferStrategy.TransferFile(source, destination);
-                        }
-                        finally
-                        {
-                            mutex.ReleaseMutex();
-                        }
-                    }
                 }
             }
         }